requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
cffi==1.10.0
cryptography==0.8
Django==1.11
django-appconf==1.0.2
django-compressor==2.1.1
django-rest-framework-nested==0.0.1
djangorestframework==3.6.2
ecdsa==0.13
enum34==1.1.6
idna==2.5
ipaddress==1.0.18
packaging==16.8
paramiko==1.15.2
pyasn1==0.2.3
pycparser==2.17
pycrypto==2.6.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
pyvirtualize==0.0.1
pyvmomi==6.5
PyYAML==3.12
rcssmin==1.0.6
requests==2.13.0
rjsmin==1.0.12
six==1.10.0
suds==0.4
xmltodict==0.11.0
